+++ NEWS.md
@@ -46,6 +46,7 @@ xxxx-xx-xx
   - GEOSBuffer: Fix crash with Inf coordinates (GH-822, Dan Baston)
   - GEOSSTRtree_iterate: Do not return removed items (GH-833, Dan Baston)
   - IndexedFacetDistance: Fix crash with Inf coordinates (GH-821, Dan Baston)
+  - HausdorffDistance: Fix crash on collection containing empty point (GH-840, Dan Baston)
 
 
 ## Changes in 3.11.0